Migrate SemVer to core
