+++ components/bridge/api/build.gradle.kts
@@ -10,6 +10,7 @@ dependencies {
     implementation(libs.protobuf.jvm)
     implementation(projects.components.core.log)
     implementation(projects.components.core.ktx)
+    implementation(projects.components.core.data)
 
     implementation(libs.kotlin.coroutines)
     implementation(libs.ble.common)
