+++ components/bridge/api/src/main/java/com/flipperdevices/bridge/api/manager/service/FlipperVersionApi.kt
@@ -1,6 +1,6 @@
 package com.flipperdevices.bridge.api.manager.service
 
-import com.flipperdevices.bridge.api.model.SemVer
+import com.flipperdevices.core.data.SemVer
 import kotlinx.coroutines.flow.StateFlow
 
 interface FlipperVersionApi {
