+++ components/bridge/impl/build.gradle.kts
@@ -10,6 +10,7 @@ dependencies {
     implementation(projects.components.core.ktx)
     implementation(projects.components.core.di)
     implementation(projects.components.core.preference)
+    implementation(projects.components.core.data)
     implementation(projects.components.analytics.shake2report.api)
     implementation(projects.components.analytics.metric.api)
