+++ components/faphub/dao/api/src/main/java/com/flipperdevices/faphub/dao/api/model/FapMetaInformation.kt
@@ -1,4 +1,7 @@
 package com.flipperdevices.faphub.dao.api.model
 
+import com.flipperdevices.core.data.SemVer
+
 data class FapMetaInformation(
+    val version: SemVer
 )
