+++ components/info/impl/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
     implementation(projects.components.core.navigation)
     implementation(projects.components.core.preference)
     implementation(projects.components.core.activityholder)
+    implementation(projects.components.core.data)
     implementation(projects.components.core.ui.res)
     implementation(projects.components.core.ui.ktx)
     implementation(projects.components.core.ui.fragment)
