+++ components/keyscreen/emulate/build.gradle.kts
@@ -13,6 +13,7 @@ dependencies {
     implementation(projects.components.core.ktx)
     implementation(projects.components.core.navigation)
     implementation(projects.components.core.preference)
+    implementation(projects.components.core.data)
     implementation(projects.components.core.ui.res)
     implementation(projects.components.core.ui.ktx)
     implementation(projects.components.core.ui.theme)
